various minor things
* debugged flash program write
* added compilation and debig settings for tests on vs code
* minor cleanups

+++ Src/rp2040/flash.c
@@ -18,7 +18,7 @@ void unlockFlash()
 {
 
 }
-uint8_t erasePage(uint8_t pagenr)
+uint8_t erasePage(uint16_t pagenr)
 {
 	flash_range_erase(pagenr << 12,4096,1,' ');
 	return 0;
@@ -29,7 +29,7 @@ uint8_t programHalfword(uint16_t hwrd,ptr addr)
 	return 0;
 }
 
-uint8_t programPage(uint8_t pagenr,uint16_t* data,uint16_t cnt)
+uint8_t programPage(uint16_t pagenr,uint16_t* data,uint16_t cnt)
 {
 	ptr fsStart = getFilesystemStart();
 	flash_range_program((uint32_t)(fsStart + (pagenr << 12)),(uint8_t*)data,cnt << 1);
